requirements.txt
phidata
groq
python-dotenv
duckduckgo-search
requests
beautifulsoup4
